some optimizations
Scratched the HMAC for a good old hash of the nonce. A 16 byte random
nonce should give us a solid 2^128 key space. Should be more than
sufficient to fend off a brute force attack. Besides HMAC with the key
exposed doesn’t really add much to the construction compared to a
simple hash